Moved layout editor to v4 and removed v3 versions
