+++ src/components/screens/floor-plan-screen.tsx
@@ -39,6 +39,7 @@ interface Props extends WithStyles<typeof styles> {
   accessToken: AccessToken;
   exhibitionId?: string;
   exhibitions: Exhibition[];
+  deviceModels: DeviceModel[];
 }
 
 /**
@@ -54,7 +55,6 @@ interface State {
   rooms: ExhibitionRoom[];
   deviceGroups: ExhibitionDeviceGroup[];
   devices: ExhibitionDevice[];
-  deviceModels: DeviceModel[];
   selectedFloor?: ExhibitionFloor;
   selectedRoom?: ExhibitionRoom;
   selectedDeviceGroup?: ExhibitionDeviceGroup;
@@ -88,7 +88,6 @@ export class FloorPlanScreen extends React.Component<Props, State> {
       rooms: [],
       deviceGroups: [],
       devices: [],
-      deviceModels: [],
       treeData: [],
       addImageDialogOpen: false
     };
@@ -211,7 +210,8 @@ export class FloorPlanScreen extends React.Component<Props, State> {
    * Renders properties
    */
   private renderProperties = () => {
-    const { cropping, cropImageDataUrl, selectedFloor, selectedRoom, selectedDeviceGroup, selectedDevice, deviceModels } = this.state;
+    const { deviceModels } = this.props;
+    const { cropping, cropImageDataUrl, selectedFloor, selectedRoom, selectedDeviceGroup, selectedDevice } = this.state;
     if (cropping && cropImageDataUrl) {
       return <FloorPlanCropProperties 
         imageHeight={ this.state.cropImageDetails?.height }
@@ -350,15 +350,13 @@ export class FloorPlanScreen extends React.Component<Props, State> {
     const exhibitionRoomsApi = Api.getExhibitionRoomsApi(accessToken);
     const exhibitionDeviceGroupsApi = Api.getExhibitionDeviceGroupsApi(accessToken);
     const exhibitionDevicesApi = Api.getExhibitionDevicesApi(accessToken);
-    const deviceModelsApi = Api.getDeviceModelsApi(accessToken);
-    const [ floors, rooms, deviceGroups, devices, deviceModels ] =
-      await Promise.all<ExhibitionFloor[], ExhibitionRoom[], ExhibitionDeviceGroup[], ExhibitionDevice[], DeviceModel[]>(
+    const [ floors, rooms, deviceGroups, devices ] =
+      await Promise.all<ExhibitionFloor[], ExhibitionRoom[], ExhibitionDeviceGroup[], ExhibitionDevice[]>(
         [
           exhibitionFloorsApi.listExhibitionFloors({ exhibitionId }),
           exhibitionRoomsApi.listExhibitionRooms({ exhibitionId }),
           exhibitionDeviceGroupsApi.listExhibitionDeviceGroups({ exhibitionId }),
-          exhibitionDevicesApi.listExhibitionDevices({ exhibitionId }),
-          deviceModelsApi.listDeviceModels()
+          exhibitionDevicesApi.listExhibitionDevices({ exhibitionId })
         ]
       );
     const selectedFloor = floors[0];
@@ -369,7 +367,6 @@ export class FloorPlanScreen extends React.Component<Props, State> {
       rooms,
       deviceGroups,
       devices,
-      deviceModels,
       selectedFloor
     });
   }
@@ -433,30 +430,30 @@ export class FloorPlanScreen extends React.Component<Props, State> {
   /**
    * Gets action buttons
    *
-   * @returns action buttons as array
+   * @returns array of action button objects
    */
-  private getActionButtons = () => {
+  private getActionButtons = (): ActionButton[] => {
     const { selectedFloor, selectedRoom, selectedDeviceGroup, selectedDevice } = this.state;
     if (selectedFloor) {
       return [
         { name: strings.generic.save, action: this.onSaveFloorClick },
         { name: strings.floorPlan.toolbar.upload, action: this.toggleUploadNewImageDialog },
         { name: strings.floorPlan.addFloor, action: this.onAddFloorClick }
-      ] as ActionButton[];
+      ];
     } else if (selectedRoom) {
       return [
         { name: strings.generic.save, action: this.onSaveRoomClick },
         { name: strings.floorPlan.addDeviceGroup, action: this.onAddDeviceGroupClick }
-      ] as ActionButton[];
+      ];
     } else if (selectedDeviceGroup) {
       return [
         { name: strings.generic.save, action: this.onSaveDeviceGroupClick },
         { name: strings.floorPlan.addDevice, action: this.onAddDeviceClick }
-      ] as ActionButton[];
+      ];
     } else if (selectedDevice) {
       return [ 
         { name: strings.generic.save, action: this.onSaveDeviceClick }
-      ] as ActionButton[];
+      ];
     } else {
       return [];
     }
@@ -464,6 +461,8 @@ export class FloorPlanScreen extends React.Component<Props, State> {
 
   /**
    * Get bounds from cropImageDetails
+   * 
+   * @returns bounds object if crop image details are found, otherwise returns undefined
    */
   private getBounds = (): Bounds | undefined => {
     const { cropImageDetails } = this.state;
@@ -644,8 +643,8 @@ export class FloorPlanScreen extends React.Component<Props, State> {
    * Event handler for add device click
    */
   private onAddDeviceClick = async () => {
-    const { accessToken, exhibitionId } = this.props;
-    const { selectedDeviceGroup, deviceModels } = this.state;
+    const { accessToken, exhibitionId, deviceModels } = this.props;
+    const { selectedDeviceGroup } = this.state;
     if (!exhibitionId || !selectedDeviceGroup || !selectedDeviceGroup.id || deviceModels.length < 1 || !deviceModels[0].id) {
       return;
     }
@@ -930,7 +929,8 @@ function mapStateToProps(state: ReduxState) {
   return {
     keycloak: state.auth.keycloak as KeycloakInstance,
     accessToken: state.auth.accessToken as AccessToken,
-    exhibitions: state.exhibitions.exhibitions
+    exhibitions: state.exhibitions.exhibitions,
+    deviceModels: state.devices.deviceModels
   };
 }
