+++ src/components/screens/layouts-screen.tsx
@@ -6,15 +6,19 @@ import { ReduxActions, ReduxState } from "../../store";
 
 import { History } from "history";
 import styles from "../../styles/exhibition-view";
-import { WithStyles, withStyles, CircularProgress } from "@material-ui/core";
+import { WithStyles, withStyles, CircularProgress, TextField, Select, MenuItem, InputLabel } from "@material-ui/core";
 import { KeycloakInstance } from "keycloak-js";
 // eslint-disable-next-line max-len
-import { PageLayout } from "../../generated/client";
+import { PageLayout, ScreenOrientation, PageLayoutViewPropertyType, DeviceModel } from "../../generated/client";
 import { AccessToken, ActionButton } from '../../types';
 import strings from "../../localization/strings";
 import CardList from "../generic/card/card-list";
 import CardItem from "../generic/card/card-item";
 import BasicLayout from "../layouts/basic-layout";
+import Api from "../../api/api";
+import { setLayouts, setSelectedLayout } from "../../actions/layouts";
+import GenericDialog from "../generic/generic-dialog";
+import theme from "../../styles/theme";
 
 /**
  * Component props
@@ -24,6 +28,9 @@ interface Props extends WithStyles<typeof styles> {
   keycloak: KeycloakInstance;
   accessToken: AccessToken;
   layouts: PageLayout[];
+  deviceModels: DeviceModel[];
+  setLayouts: typeof setLayouts;
+  setSelectedLayout: typeof setSelectedLayout;
 }
 
 /**
@@ -31,6 +38,8 @@ interface Props extends WithStyles<typeof styles> {
  */
 interface State {
   loading: boolean;
+  addNewDialogOpen: boolean;
+  newLayout: Partial<PageLayout>;
 }
 
 /**
@@ -46,7 +55,9 @@ class LayoutsScreen extends React.Component<Props, State> {
   constructor(props: Props) {
     super(props);
     this.state = {
-      loading: false
+      loading: false,
+      addNewDialogOpen: false,
+      newLayout: {}
     };
   }
 
@@ -55,7 +66,6 @@ class LayoutsScreen extends React.Component<Props, State> {
    */
   public render = () => {
     const { classes, history, keycloak } = this.props;
-    const actionBarButtons = this.getActionButtons();
 
     if (this.state.loading) {
       return (
@@ -64,7 +74,7 @@ class LayoutsScreen extends React.Component<Props, State> {
           history={ history }
           title={ strings.layout.title }
           breadcrumbs={ [] }
-          actionBarButtons={ actionBarButtons }
+          actionBarButtons={ [] }
           noBackButton
         >
           <div className={ classes.loader }>
@@ -80,10 +90,11 @@ class LayoutsScreen extends React.Component<Props, State> {
         history={ history }
         title={ strings.layout.title }
         breadcrumbs={ [] }
-        actionBarButtons={ actionBarButtons }
+        actionBarButtons={ this.getActionButtons() }
         noBackButton
       >
         { this.renderLayoutCardsList() }
+        { this.renderAddNewDialog() }
       </BasicLayout>
     );
   }
@@ -93,6 +104,10 @@ class LayoutsScreen extends React.Component<Props, State> {
    */
   private renderLayoutCardsList = () => {
     const { layouts } = this.props;
+    if (layouts.length < 1) {
+      return null;
+    }
+    
     const cardMenuOptions = this.getCardMenuOptions();
     const cards = layouts.map(exhibition => {
       const layoutId = exhibition.id;
@@ -119,6 +134,53 @@ class LayoutsScreen extends React.Component<Props, State> {
   }
 
   /**
+   * Renders add new dialog
+   */
+  private renderAddNewDialog = () => {
+    const { deviceModels } = this.props;
+    const { newLayout, addNewDialogOpen } = this.state;
+    return (
+      <GenericDialog
+        error={ !newLayout.name }
+        open={ addNewDialogOpen }
+        title={ strings.layout.addNew }
+        positiveButtonText={ strings.generic.add }
+        cancelButtonText={ strings.generic.cancel }
+        onCancel={ this.toggleAddNewDialog }
+        onClose={ this.toggleAddNewDialog }
+        onConfirm={ this.createNewLayout }
+      >
+        <TextField
+          fullWidth
+          variant="outlined"
+          label={ strings.generic.name }
+          name="name"
+          value={ newLayout.name }
+          onChange={ this.onNewLayoutChange }
+        />
+        <InputLabel id="screenOrientation-label" style={{ marginTop: theme.spacing(2) }}>
+          { strings.layout.settings.deviceModelId }
+        </InputLabel>
+        <Select
+          fullWidth
+          variant="filled"
+          style={{ marginTop: theme.spacing(2) }}
+          label={ strings.dashboard.devices.dialog.model }
+          name="modelId"
+          value={ newLayout.modelId || "" }
+          onChange={ this.onNewLayoutChange }
+        >
+          { deviceModels.map(model => 
+            <MenuItem key={ model.id } value={ model.id }>
+              { `${model.manufacturer} ${model.model}` }
+            </MenuItem>
+          )}
+        </Select>
+      </GenericDialog>
+    );
+  }
+
+  /**
    * Gets card menu options
    *
    * @returns card menu options as action button array
@@ -137,16 +199,89 @@ class LayoutsScreen extends React.Component<Props, State> {
    */
   private getActionButtons = () => {
     return [
-      { name: "add layout", action: () => null }
+      { name: strings.layout.addNew, action: this.toggleAddNewDialog }
     ] as ActionButton[];
   }
 
   /**
+   * Creates new layout
+   * 
+   * @param layout layout
+   */
+  private createNewLayout = async () => {
+    const { accessToken, setLayouts, layouts } = this.props;
+    const { newLayout } = this.state;
+    if (!newLayout.name || !newLayout.modelId) {
+      return;
+    }
+
+    const pageLayout: PageLayout = {
+      name: newLayout.name,
+      screenOrientation: ScreenOrientation.Landscape,
+      modelId: newLayout.modelId,
+      data: {
+        id: "newFrameLayout",
+        widget: "FrameLayout",
+        properties: [{
+          name: "layout_width",
+          value: "match_parent",
+          type: PageLayoutViewPropertyType.String
+        }, {
+          name: "layout_height",
+          value: "match_parent",
+          type: PageLayoutViewPropertyType.String
+        }, {
+          name: "background",
+          value: "#000000",
+          type: PageLayoutViewPropertyType.String
+        }],
+        children: []
+      }
+    };
+    const layoutsApi = Api.getPageLayoutsApi(accessToken);
+    const createdLayout = await layoutsApi.createPageLayout({ pageLayout });
+    layouts.push(createdLayout);
+    setLayouts(layouts);
+  }
+
+  /**
    * Event handler for card click
+   * 
+   * @param layoutId layout id
    */
   private onCardClick = (layoutId: string) => {
+    const { layouts } = this.props;
     const { pathname } = this.props.history.location;
+    const layout = layouts.find(layout => layout.id === layoutId);
+    if (!layout) {
+      return;
+    }
+
+    this.props.setSelectedLayout(layout);
     this.props.history.push(`${pathname}/${layoutId}`);
+  }
+
+  /**
+   * Toggles add new layout dialog
+   */
+  private toggleAddNewDialog = () => {
+    this.setState({
+      addNewDialogOpen: !this.state.addNewDialogOpen
+    });
+  }
+
+  /**
+   * Event handler for new layout change
+   * 
+   * @param event event
+   */
+  private onNewLayoutChange = (event: React.ChangeEvent<HTMLInputElement | { name?: string; value: any }>) => {
+    const { newLayout } = this.state;
+    const { name, value } = event.target;
+    if (!name) {
+      return;
+    }
+    this.setState({ newLayout: { ...newLayout, [name]: value } });
   }
 
   /**
@@ -167,7 +302,8 @@ function mapStateToProps(state: ReduxState) {
   return {
     keycloak: state.auth.keycloak as KeycloakInstance,
     accessToken: state.auth.accessToken as AccessToken,
-    layouts: state.layouts.layouts
+    layouts: state.layouts.layouts,
+    deviceModels: state.devices.deviceModels
   };
 }
 
@@ -178,6 +314,8 @@ function mapStateToProps(state: ReduxState) {
  */
 function mapDispatchToProps(dispatch: Dispatch<ReduxActions>) {
   return {
+    setLayouts: (layouts: PageLayout[]) => dispatch(setLayouts(layouts)),
+    setSelectedLayout: (layout: PageLayout) => dispatch(setSelectedLayout(layout))
   };
 }
